+++ ic-agent/src/lib.rs
@@ -68,11 +68,6 @@
 //!   agent.fetch_root_key().await?;
 //!   let management_canister_id = Principal::from_text("aaaaa-aa")?;
 //!
-//!   let waiter = garcon::Delay::builder()
-//!     .throttle(std::time::Duration::from_millis(500))
-//!     .timeout(std::time::Duration::from_secs(60 * 5))
-//!     .build();
-//!
 //!   // Create a call to the management canister to create a new canister ID,
 //!   // and wait for a result.
 //!   // The effective canister id must belong to the canister ranges of the subnet at which the canister is created.
@@ -80,7 +75,7 @@
 //!   let response = agent.update(&management_canister_id, "provisional_create_canister_with_cycles")
 //!     .with_effective_canister_id(effective_canister_id)
 //!     .with_arg(&Encode!(&Argument { amount: None})?)
-//!     .call_and_wait(waiter)
+//!     .call_and_wait()
 //!     .await?;
 //!
 //!   let result = Decode!(response.as_slice(), CreateCanisterResult)?;